Refactor logging package structure and move context.go to kopialib/backend Signed-off-by: Priyansh Choudhary <im1706@gmail.com>
